Add directory source code for publisher
* Move auto-publish.py -> src/ign-publisher/auto-publish.py
* Move publish-converted-image.sh -> src/ign-publisher/publish-converted-image.sh
* Refactoring auto-publish.py
* Move Dockerfile -> src/ign-publisher/Dockerfile
* Update dockerfile with new src of ign-publisher
