--- publish-converted-image.sh
@@ -1,3 +1,0 @@
-#!/bin/sh -l
-
-python auto-publish.py
