+++ src/ign-publisher/publish-converted-image.sh
@@ -1,0 +1,3 @@
+#!/bin/sh -l
+
+python auto-publish.py
